refactor: split OpenAI overload retry helpers
