+++ .agents/plugins/opencode-aidevops/openai-provider-auth.mjs
@@ -7,22 +7,13 @@
  */
 
 import { getAccounts, patchAccount, rotateOpenAIPoolToken } from "./oauth-pool.mjs";
+import { isOpenAIOverloadText, overloadRetryDelaysMs, sleep, wrapOpenAIOverloadStream } from "./openai-overload-retry.mjs";
+
+export { isOpenAIOverloadText } from "./openai-overload-retry.mjs";
 
 const OPENAI_API_HOST = "api.openai.com";
 const OPENAI_API_PREFIX = "/v1/";
 const DEFAULT_OPENAI_COOLDOWN_MS = 300_000;
-const DEFAULT_OVERLOAD_RETRY_DELAYS_MS = [2_000, 5_000, 10_000];
-const OVERLOAD_MARKERS = [
-  "service_unavailable_error",
-  "server_is_overloaded",
-  "servers are currently overloaded",
-];
-const STREAM_CONTENT_MARKERS = [
-  "response.output_text.delta",
-  "response.function_call_arguments.delta",
-  "response.code_interpreter_call_code.delta",
-  "response.reasoning_summary_text.delta",
-];
 const USAGE_LIMIT_MARKERS = [
   "usage limit",
   "rate limit",
@@ -42,28 +33,6 @@ export function parseRetryAfterMs(response) {
   const date = Date.parse(raw);
   if (Number.isFinite(date)) return Math.max(date - Date.now(), DEFAULT_OPENAI_COOLDOWN_MS);
   return DEFAULT_OPENAI_COOLDOWN_MS;
-}
-
-function overloadRetryDelaysMs() {
-  const raw = process.env.AIDEVOPS_OPENAI_OVERLOAD_RETRY_DELAYS_MS || "";
-  const parsed = raw
-    .split(",")
-    .map((part) => Number.parseInt(part.trim(), 10))
-    .filter((value) => Number.isFinite(value) && value >= 0);
-  return parsed.length > 0 ? parsed : DEFAULT_OVERLOAD_RETRY_DELAYS_MS;
-}
-
-function sleep(ms) {
-  return new Promise((resolve) => setTimeout(resolve, ms));
-}
-
-export function isOpenAIOverloadText(text) {
-  const lowered = String(text || "").toLowerCase();
-  return OVERLOAD_MARKERS.some((marker) => lowered.includes(marker));
-}
-
-function isOpenAIStreamContentText(text) {
-  return STREAM_CONTENT_MARKERS.some((marker) => String(text || "").includes(marker));
 }
 
 function requestUrl(input) {
@@ -175,100 +144,6 @@ async function handleOpenAIOverload(ctx) {
   return lastResponse;
 }
 
-function enqueueChunks(controller, chunks) {
-  for (const chunk of chunks) controller.enqueue(chunk);
-}
-
-async function pipeRemainingStream(reader, controller) {
-  while (true) {
-    const { done, value } = await reader.read();
-    if (done) {
-      controller.close();
-      return;
-    }
-    controller.enqueue(value);
-  }
-}
-
-async function inspectStreamPrefix(reader, controller, decoder, retryAvailable) {
-  const buffered = [];
-  let bufferedText = "";
-
-  while (true) {
-    const { done, value } = await reader.read();
-    if (done) {
-      enqueueChunks(controller, buffered);
-      controller.close();
-      return "closed";
-    }
-
-    const text = decoder.decode(value, { stream: true });
-    buffered.push(value);
-    bufferedText += text;
-
-    if (isOpenAIOverloadText(bufferedText)) {
-      if (retryAvailable) {
-        await reader.cancel().catch(() => {});
-        return "retry";
-      }
-      enqueueChunks(controller, buffered);
-      return "pipe";
-    }
-
-    if (isOpenAIStreamContentText(bufferedText) || bufferedText.length > 65_536) {
-      enqueueChunks(controller, buffered);
-      return "pipe";
-    }
-  }
-}
-
-async function retryOpenAIOverloadStream(ctx) {
-  const { originalFetch, response, retryInput, init, controller, retryDelays } = ctx;
-  const decoder = new TextDecoder();
-  let currentResponse = response;
-
-  for (let attempt = 0; ; attempt += 1) {
-    const reader = currentResponse.body.getReader();
-    const outcome = await inspectStreamPrefix(reader, controller, decoder, attempt < retryDelays.length);
-    if (outcome === "closed") return;
-    if (outcome === "pipe") return pipeRemainingStream(reader, controller);
-
-    const delayMs = retryDelays[attempt];
-    console.error(`[aidevops] OpenAI provider: overloaded stream error — retrying request (${attempt + 1}/${retryDelays.length})`);
-    if (delayMs > 0) await sleep(delayMs);
-    currentResponse = await originalFetch(buildRetryRequest(retryInput), init);
-    if (!currentResponse.body) {
-      controller.close();
-      return;
-    }
-  }
-}
-
-function wrapOpenAIOverloadStream(ctx) {
-  const { originalFetch, response, retryInput, init } = ctx;
-  if (!response.body) return response;
-  const contentType = response.headers?.get?.("content-type") || "";
-  if (!contentType.includes("text/event-stream")) return response;
-
-  const retryDelays = overloadRetryDelaysMs();
-
-  const stream = new ReadableStream({
-    async start(controller) {
-      try {
-        await retryOpenAIOverloadStream({ originalFetch, response, retryInput, init, controller, retryDelays });
-      } catch (err) {
-        controller.error(err);
-      }
-    },
-  });
-
-  return new Response(stream, {
-    status: response.status,
-    statusText: response.statusText,
-    headers: response.headers,
-  });
-}
-
 async function maybeRotateBeforeOpenAIFetch(client, input, init) {
   const current = resolveOpenAIAccount(extractBearerToken(input, init));
   if (!isUnavailableAccount(current)) return init;
@@ -296,7 +171,7 @@ export function installOpenAIProviderFetchRotation(client) {
       return handleOpenAIOverload({ originalFetch, init: firstInit, response, retryInput });
     }
     if (response.ok) {
-      return wrapOpenAIOverloadStream({ originalFetch, response, retryInput, init: firstInit });
+      return wrapOpenAIOverloadStream({ originalFetch, response, retryInput, init: firstInit, buildRetryRequest });
     }
     return response;
   };
